Fixed get_epncore_meta() method

+++ src/maser_data/maser/data/base/base.py
@@ -196,8 +196,8 @@ class Data(BaseData, dataset="default"):
         :return md: a dict with time_min, time_max and granule_gid (from dataset_name attribute) keys.
         """
         md = dict()
-        md["time_min"] = self.times[0].jd
-        md["time_max"] = self.times[-1].jd
+        md["time_min"] = self.times[0].jd.astype(float)
+        md["time_max"] = self.times[-1].jd.astype(float)
         md["granule_gid"] = self.dataset
         md["file_name"] = self.filepath.name
         md["access_format"] = self.mime_type
